added delete reply repository changes

+++ src/main/java/com/tweetapp/repositories/TweetReplyRepository.java
@@ -12,6 +12,6 @@ public interface TweetReplyRepository extends MongoRepository<TweetReply, Long> 
 
 	List<TweetReply> findByParenttweetId(long parenttweetId);
 
-	public long deleteByparenttweetId(long parenttweetId);
+	public int deleteByparenttweetId(long parenttweetId);
 
 }
